refactor: setup screens
