+++ apps/guardian-ui/src/components/setup/screens/connectGuardians/ConnectGuardians.tsx
@@ -17,11 +17,11 @@ import {
 import { CopyInput, Table, TableRow } from '@fedimint/ui';
 import { ModuleKind, ServerStatus } from '@fedimint/types';
 import { useTranslation } from '@fedimint/utils';
-import { useConsensusPolling, useSetupContext } from '../../hooks';
-import { GuardianRole } from '../../types';
-import { getModuleParamsFromConfig } from '../../utils/api';
-import { ReactComponent as CopyIcon } from '../../assets/svgs/copy.svg';
-import { BftInfo } from '../BftInfo';
+import { useConsensusPolling, useSetupContext } from '../../../../hooks';
+import { GuardianRole } from '../../../../types';
+import { getModuleParamsFromConfig } from '../../../../utils/api';
+import { ReactComponent as CopyIcon } from '../assets/svgs/copy.svg';
+import { BftInfo } from '../../../BftInfo';
 
 interface Props {
   next(): void;
