+++ apps/guardian-ui/src/components/setup/screens/roleSelector/RoleSelector.tsx
@@ -10,14 +10,14 @@ import {
   Flex,
 } from '@chakra-ui/react';
 import { RadioButtonGroup, RadioButtonOption } from '@fedimint/ui';
-import { GuardianRole, SETUP_ACTION_TYPE } from '../../types';
-import { ReactComponent as ArrowRightIcon } from '../../assets/svgs/arrow-right.svg';
-import { ReactComponent as CheckIcon } from '../../assets/svgs/check.svg';
-import { ReactComponent as StarsIcon } from '../../assets/svgs/stars.svg';
-import { ReactComponent as IntersectSquareIcon } from '../../assets/svgs/intersect-square.svg';
-import { ReactComponent as WarningIcon } from '../../assets/svgs/warning.svg';
-import { ReactComponent as SoloIcon } from '../../assets/svgs/solo.svg';
-import { useSetupContext } from '../../hooks';
+import { GuardianRole, SETUP_ACTION_TYPE } from '../../../../types';
+import { ReactComponent as ArrowRightIcon } from '../assets/svgs/arrow-right.svg';
+import { ReactComponent as CheckIcon } from '../assets/svgs/check.svg';
+import { ReactComponent as StarsIcon } from '../assets/svgs/stars.svg';
+import { ReactComponent as IntersectSquareIcon } from '../assets/svgs/intersect-square.svg';
+import { ReactComponent as WarningIcon } from '../assets/svgs/warning.svg';
+import { ReactComponent as SoloIcon } from '../assets/svgs/solo.svg';
+import { useSetupContext } from '../../../../hooks';
 import { useTranslation } from '@fedimint/utils';
 
 interface Props {
